feat(ui-kit): port state-views.tsx primitives into @loopover/ui-kit
Spinner, LoadingState, EmptyState, StateActionButton, ErrorState, and
StateBoundary move to packages/loopover-ui-kit/src/components/state-views.tsx
so apps/loopover-miner-ui (and the future chat message list, per #6244's
audit) can consume them directly instead of reaching into apps/loopover-ui.

apps/loopover-ui's own state-views.tsx becomes a back-compat wrapper:
Spinner/LoadingState/EmptyState/StateActionButton/ErrorState re-export
unchanged, StateBoundary forwards every prop and defaults onFailureNotify
to the app's real notifyApiFailure, and usePreviewDataState (unrelated to
this move, zero call sites) stays put. Every existing call site's runtime
behavior is unchanged -- confirmed by the app's own state-views.test.tsx
(14/14 passing with zero edits) and the broader site/routes suite
(230/230 passing).

The ui-kit copy can't import from apps/loopover-ui, so it defines its own
ApiFailureKind (structurally identical to request.ts's) and replaces the
hardcoded notifyApiFailure call with an optional onFailureNotify prop that
no-ops when the caller doesn't supply one -- the wrapper passes it as a
lazily-dereferenced closure (not a direct reference) so tests that
partially mock @/lib/api/request without notifyApiFailure, and never
exercise the error path, keep working exactly as before.

Closes #6506
